Superstar2 can now stash raw measurements -
not that it's in any way useful yet...

+++ gps.h
@@ -899,7 +899,7 @@ struct gps_data_t {
 
 #ifdef RAW_ENABLE
     /* raw measurement data */
-    float codephase[MAXCHANNELS];	/* meters */
+    double codephase[MAXCHANNELS];	/* meters */
     double carrierphase[MAXCHANNELS];	/* meters */
     double pseudorange[MAXCHANNELS];	/* meters */
     double deltarange[MAXCHANNELS];	/* meters/sec */
